frameworks/personal-companion/requirements.txt
playwright
undetected-playwright
openai
watchdog
python-dotenv
google-generativeai
beautifulsoup4
lxml